+++ font/rasterizer.h
@@ -12,7 +12,7 @@ struct RasterizedGlyph {
     int32_t top;
     int32_t width;
     int32_t height;
-    float advance;
+    int32_t advance;
     std::vector<uint8_t> buffer;
     // TODO: Either remove these debug fields, or evaluate if we should keep them.
     unsigned short index;
